backend/quizzler/requirements.txt
asgiref==3.8.1
channels==4.2.2
Django==5.2
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
PyJWT==2.9.0
PyMySQL==1.1.1
python-dotenv==1.1.0
sqlparse==0.5.3
